Add Actions resolver support (#1)
## Description
Adds support for the actions resolver in the GraphQL server. This PR has
the following changes:
1. Some refactoring to move the used SQL into its module and then
defines a separate CTE for fetching actions.
2. Adds additional code to the `ArchiveNodeAdapter` class to use the
newly defined SQL
3. Fixes some issues with the schema when actions still needed to be
implemented. The schema was just defined incorrectly for actions
previously.

## Screenshots
The screenshot below shows a GraphiQL environment that issues a
`actions` query, and it's output.

![image](https://user-images.githubusercontent.com/9512405/218589442-e47d936e-c09a-485c-a656-78aef8ea6611.png)

## Impact
This allows querying for actions on the current running Berkeley
network. One important note is that the current Berkeley archive node
uses an older schema that references actions as `sequence_events`, so
the SQL must be updated in a follow-up PR when Berkeley is redeployed.

## Testing
Manually tested using a [Berkely Archive Node dump as of February
13th](https://storage.googleapis.com/mina-archive-dumps/berkeley-archive-dump-2023-02-13_0000.sql.tar.gz.