feat: Add full-chain CTE
Adds a new CTE that gets the full chain of blocks. This is now used in
the blocks CTE, with the idea that instead of selecting all blocks, we
select the blocks that form a path from 0 to the current max height.
There can be blocks submitted from block producers that do not make it
into the chain and we want to ignore such blocks.