Add full-chain CTE (#6)
## Description
Adds a new CTE that gets the full chain of blocks. This is now used in
the blocks CTE, with the idea that instead of selecting all blocks, we
select the blocks that form a path from 0 to the current max height.

## Impact
There can be blocks submitted from block producers that do not make it
into the chain, and we want to ignore them. By using this new CTE, we
only use the blocks that form the expected chain and ignore
events/actions in these duplicated blocks.

## Tested
Manual testing