feat: Add distanceFromMaximumHeight field on blockInfo
Adds a new field to the blockInfo GraphQL type. As the name suggests, it
is the distance from the maximum block height, for a given block. This
will be useful for when a zkApp developer is trying to query for
event/action data at the very tip of the network. When the distance is
0, there can be multiple blocks returned as there is no good selection
algorithm implemented to decide which block to return at the very tip.