Add distance from max block height  (#13)
## Description
Adds a new field to the blockInfo GraphQL type. As the name suggests, it
is the distance from the maximum block height for a given block. This
will be useful for when a zkApp developer is trying to query for
event/action data at the very tip of the network. When the distance is
0, multiple blocks can be returned as there is no good selection
algorithm implemented to decide which block to return at the very tip.

Also, change some String values on the GraphQL schema to numbers since
they make more sense. Additionally, make a name change to the 'fields'
field on Events to be 'data' instead.

Adds a context file which exports a GraphQLContext interface and a
buildContext function. Before this change, the context is passed into
the resolver was being typed as `any`. This change types the context to
have better type safety in our resolvers.

## Impact
Now have a GraphQL field to indicate when the event/actions data is at
the current tip of the network. Additionally, there is better type of
safety in the project now.

## Testing
Manually tested

![image](https://user-images.githubusercontent.com/9512405/219231192-c18dface-e69d-45dd-9e74-998ecaf72bb3.png)