Add example `.env` file (#21)
## Description

This PR adds an example `.env.example` file to the project. This file
serves as a reference for setting up environment variables required by
the TypeScript server. The file contains a list of default values for
each environment variable, which can be used as a guide to set up the
actual `.env` file.

## Impact

Adding the `.env.example` file will make it easier for developers to set
up the environment variables required for the TypeScript server. They
can simply copy the contents of the .env.example file and use it as a
starting point to create their own `.env` file. This will reduce the
risk of misconfigured environment variables and ensure the server is set
up and running smoothly. Additionally, this file will provide a clear
and concise reference for the environment variables required by the
server, making it easier for new developers to get started with the
project.

Addresses https://github.com/o1-labs/Archive-Node-API/issues/15