Add tests (#26)
## Description
This PR introduces testing to the TypeScript GraphQL server, using a
test runner called Vitest. Vitest is a fast and user-friendly
alternative to Jest, and offers improvements in developer experience.

Two types of tests are introduced: integration tests and unit tests. The
integration tests use Github Actions to spin up a mocked GraphQL server
and a Postgres container in CI, using a nightly Berkeley dump. The
nightly Berkeley dump is automatically downloaded with the current date
specified. These tests verify that the resolvers can be called correctly
with their defined queries.

The unit tests for the ArchiveNodeAdapter class use mocked data from the
Postgres database to verify its correctness. The mocked data is
generated with a new script that simply connects to a running Postgres
database and logs some rows corresponding to a specified public key.
These are also run in CI.

This PR also adds a script to download a nightly Berkeley dump with the
specified date. This can be used to easily pull a database dump to spin
up a local Postgres instance easily and is used in CI as well.

## Impact
The addition of these tests provides a reliable and repeatable way to
verify the correctness of the GraphQL server and its components. This
will lead to increased confidence in new features, bug fixes, and better
overall application stability. Vitest also provides a faster and more
user-friendly testing experience for developers.