Use .graphql extension for schema instead of .ts (#48)
## Description

This PR switches the GraphQL schema format from .ts to .graphql. With
this change, we can now define our schema in a more concise and
intuitive format.

## Impact

This change improves the overall maintainability of our application by
providing a more standardized and portable way to define our GraphQL
schema. It also reduces the boilerplate code required to define our
schema, making it easier to reason about and debug. This change also
simplifies the build process by reducing the number of TypeScript files
that need to be compiled, resulting in faster build times. Overall, this
change streamlines our development process and makes it easier for new
developers to get up and running with our GraphQL schema.