Add graphql inspector yaml (#40)
## Description

Adds [GraphQL Inspector](https://the-guild.dev/graphql/inspector) as a
tool to be run in CI that checks for breaking changes in the GraphQL
schema. If the CI finds a breaking change, it will automatically fail
the PR and require an `expected-breaking-change` label to be applied.

This forces a review to occur if there is a breaking change and reduces
the chance of something sneaking in accidentally.