Remove index from GraphQL schema (#55)
## Description

If a zkApp only stores one event type, an index is not stored in the
Archive Node to save space. With the possibility of a zkApp having an
index or not having an index, it's simpler just to remove the index
property and use the data property to hold an index if it's available.
The logic in SnarkyJS already does this assumption where if there is
only a single index type, it will treat the data property as only the
field element.