Fix duplicate actions and events of same type being emitted (#64)
## Description

Fixes the following bugs:

1. Actions were being duplicated, similar to events. The same fix for
events was applied for actions: filtering blocks before processing them
for the action data.
2. zkApps that emitted multiple events with the same value multiple
times in the same Account Update needed to be reported correctly. The
API only showed them once instead of showing all the emitted events with
the same data.

Additionally, an optimization was done to reduce the number of columns
that were being selected. Now, only the set of data that is needed is
being queried.