feat(run-compare.ts): add new file to compare TypeScript and OCaml files
This new file, run-compare.ts, is created to compare the contents of TypeScript and OCaml files. It checks if the directories contain the same number of files, if the file names match, and if the file contents are equal. This is done to ensure consistency between TypeScript and OCaml files.