adjust query to support ordering (#107)
Fixes: #106 

This PR adds `sequenceNumber`, as well as `zkappAccountUpdateIds` to the
graphQL schema, which will expose enough information for clients to
verify the order in which actions were processed by consensus. It also
attempts to return actions in that order, and exposes the ordering spec
in a unit test.

We also overhaul the test contract and test suite to emit more random
and different values so that sequential events and actions can be
identified by their data.

I have tested this branch locally against the [sibling
branch](https://github.com/o1-labs/o1js/pull/1917) in o1js.