move pending blocks query out of recursive CTE (#116)
Follow on to #115 .....

This PR moves the query our of the recursive CTE and into the full chain
union query. The height referenced in the recursive CTE is not usable as
I was trying previously.

Example screenshots from the old code:

As you can see, it is possible to query the right event, but the height
filter is not the expected number.


![image](https://github.com/user-attachments/assets/6132de6f-79d9-4607-9a0c-9918ecc5bda7)


![image](https://github.com/user-attachments/assets/2689338b-d5c4-494f-8005-e7a84b08f9b2)

New code:

As you see, the specific query now selects the correct actions.


![image](https://github.com/user-attachments/assets/5b307d30-2b3e-42e7-af1f-917e6fed700b)